feat: separate createRoute and columnTypeConstants
