+++ src/scripts/modules/configurations/react/components/ConfigurationRowsTableCell.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import immutableMixin from 'react-immutable-render-mixin';
 import StorageApiTableLinkEx from '../../../components/react/components/StorageApiTableLinkEx';
-import {columnTypes} from '../../../configurations/utils/createRoute';
+import columnTypes from '../../../configurations/utils/columnTypeConstants';
 import StorageApiBucketLink from '../../../components/react/components/StorageApiBucketLink';
 import StorageApiFileUploadsLink from '../../../components/react/components/StorageApiFileUploadsLink';
