+++ src/scripts/modules/configurations/utils/columnTypeConstants.js
@@ -1,0 +1,6 @@
+const columnTypeConstants = {
+  STORAGE_LINK_DEFAULT_BUCKET: 'storage-link-default-bucket',
+  VALUE: 'value'
+};
+
+export default columnTypeConstants;
