+++ src/scripts/modules/configurations/utils/createRoute.js
@@ -11,12 +11,8 @@ import InstalledComponentsStore from '../../components/stores/InstalledComponent
 import ConfigurationRowsStore from '../ConfigurationRowsStore';
 import _ from 'lodash';
 import fuzzy from 'fuzzy';
-var Immutable = require('immutable');
-
-const columnTypeConstants = {
-  STORAGE_LINK_DEFAULT_BUCKET: 'storage-link-default-bucket',
-  VALUE: 'value'
-};
+import Immutable from 'immutable';
+import columnTypeConstants from './columnTypeConstants';
 
 // defaults
 const defaults = {
@@ -72,59 +68,54 @@ const defaults = {
   }
 };
 
-module.exports = {
-  columnTypes: columnTypeConstants,
-  createRoute(settings) {
-    const settingsWithDefaults = _.defaultsDeep(settings, defaults);
-    let route = {
-      name: settingsWithDefaults.componentId,
+export default function(settings) {
+  const settingsWithDefaults = _.defaultsDeep(settings, defaults);
+  let route = {
+    name: settingsWithDefaults.componentId,
+    settings: settingsWithDefaults,
+    path: ':config',
+    title: (routerState) => {
+      const configId = routerState.getIn(['params', 'config']);
+      return InstalledComponentsStore.getConfig(settingsWithDefaults.componentId, configId).get('name');
+    },
+    isComponent: true,
+    defaultRouteHandler: Index,
+    poll: {
+      interval: 10,
+      action: (params) => jobsActions.loadComponentConfigurationLatestJobs(settingsWithDefaults.componentId, params.config)
+    },
+    requireData: [
+      (params) => installedComponentsActions.loadComponentConfigData(settingsWithDefaults.componentId, params.config),
+      (params) => versionsActions.loadVersions(settingsWithDefaults.componentId, params.config)
+    ],
+    childRoutes: []
+  };
+  route.childRoutes.push(
+    {
+      name: settingsWithDefaults.componentId + '-row',
       settings: settingsWithDefaults,
-      path: ':config',
+      path: 'rows/:row',
       title: (routerState) => {
         const configId = routerState.getIn(['params', 'config']);
-        return InstalledComponentsStore.getConfig(settingsWithDefaults.componentId, configId).get('name');
-      },
-      isComponent: true,
-      defaultRouteHandler: Index,
-      poll: {
-        interval: 10,
-        action: (params) => jobsActions.loadComponentConfigurationLatestJobs(settingsWithDefaults.componentId, params.config)
+        const rowId = routerState.getIn(['params', 'row']);
+        const configurationRow = ConfigurationRowsStore.get(settingsWithDefaults.componentId, configId, rowId);
+        return configurationRow.get('name') !== '' ? configurationRow.get('name') : 'Untitled ' + settingsWithDefaults.row.name.singular;
       },
       requireData: [
-        (params) => installedComponentsActions.loadComponentConfigData(settingsWithDefaults.componentId, params.config),
-        (params) => versionsActions.loadVersions(settingsWithDefaults.componentId, params.config)
+        (params) => rowVersionsActions.loadVersions(settingsWithDefaults.componentId, params.config, params.row),
+        () => storageActions.loadTables()
       ],
-      childRoutes: []
-    };
-    route.childRoutes.push(
-      {
-        name: settingsWithDefaults.componentId + '-row',
-        settings: settingsWithDefaults,
-        path: 'rows/:row',
-        title: (routerState) => {
-          const configId = routerState.getIn(['params', 'config']);
-          const rowId = routerState.getIn(['params', 'row']);
-          const configurationRow = ConfigurationRowsStore.get(settingsWithDefaults.componentId, configId, rowId);
-          return configurationRow.get('name') !== '' ? configurationRow.get('name') : 'Untitled ' + settingsWithDefaults.row.name.singular;
-        },
-        requireData: [
-          (params) => rowVersionsActions.loadVersions(settingsWithDefaults.componentId, params.config, params.row),
-          () => storageActions.loadTables()
-
-        ],
-
-        defaultRouteHandler: Row,
-        childRoutes: [
-          {
-            name: settingsWithDefaults.componentId + '-row-versions',
-            settings: settingsWithDefaults,
-            path: 'versions',
-            title: 'Versions',
-            defaultRouteHandler: Versions
-          }
-        ]
-      }
-    );
-    return route;
-  }
-};
+      defaultRouteHandler: Row,
+      childRoutes: [
+        {
+          name: settingsWithDefaults.componentId + '-row-versions',
+          settings: settingsWithDefaults,
+          path: 'versions',
+          title: 'Versions',
+          defaultRouteHandler: Versions
+        }
+      ]
+    }
+  );
+  return route;
+}
